Reintroduce make_image_reader to package
