+++ sensenet/preprocess/image.py
@@ -112,6 +112,27 @@ def rescale(settings, target_shape, image):
     else:
         return new_image
 
+def make_image_reader(settings, target_shape):
+    n_chan = target_shape[-1]
+    input_format = settings.input_image_format or 'file'
+    prefix = settings.image_path_prefix or '.'
+
+    def read_image(path_or_bytes):
+        if input_format == 'pixel_values':
+            raw = path_or_bytes
+        else:
+            if input_format == 'image_bytes':
+                img_bytes = path_or_bytes
+            else:
+                path = tf.strings.join([prefix + os.sep, path_or_bytes])
+                img_bytes = tf.io.read_file(path)
+
+            raw = tf.io.decode_jpeg(img_bytes, dct_method=DCT, channels=n_chan)
+
+        return rescale(settings, target_shape, raw)
+
+    return read_image
+
 class ImageReaderLayer(tf.keras.layers.Layer):
     def __init__(self, **kwargs):
         layer_args = dict(kwargs)
