+++ tests/utils.py
@@ -17,24 +17,3 @@ def read_regression(path, root=TEST_DATA_DIR):
 def read_zipped_json(path):
     with gzip.open(path, "rb") as fin:
         return json.loads(fin.read().decode('utf-8'))
-
-def make_image_reader(settings, target_shape):
-    n_chan = target_shape[-1]
-    input_format = settings.input_image_format or 'file'
-    prefix = settings.image_path_prefix or '.'
-
-    def read_image(path_or_bytes):
-        if input_format == 'pixel_values':
-            raw = path_or_bytes
-        else:
-            if input_format == 'image_bytes':
-                img_bytes = path_or_bytes
-            else:
-                path = tf.strings.join([prefix + os.sep, path_or_bytes])
-                img_bytes = tf.io.read_file(path)
-
-            raw = tf.io.decode_jpeg(img_bytes, dct_method=DCT, channels=n_chan)
-
-        return rescale(settings, target_shape, raw)
-
-    return read_image
